app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "langchain>=0.3.18",
    "langchain-community>=0.3.17",
    "langchain-openai>=0.3.4",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.2.70",
    "pymupdf>=1.25.3",
    "pypdf>=5.3.0",
    "python-dotenv>=1.0.1",
]
